requirements.txt
absl-py==0.9.0
afb==1.3.1.2
alembic==1.4.0
certifi==2019.11.28
Click==7.0
ecdsa==0.15
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.8.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
pycryptodome==3.9.7
python-dateutil==2.8.1
python-editor==1.0.4
python-jose-cryptodome==1.3.2
PyYAML==5.3
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
